docker/telegram-bot/requirements-telegram_bot.txt
requests>=2.28
python-telegram-bot>=20.0